requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
streamlit
python-dotenv